+++ climpred/bootstrap.py
@@ -376,16 +376,18 @@ def bootstrap_uninit_pm_ensemble_from_control_cftime(init_pm, control):
 
 def resample_uninitialized_from_initialized(init, resample_dim=["init", "member"]):
     """
-    Generate an uninitialized ensemble by resampling from the initialized prediction ensemble.
+    Generate an uninitialized ensemble by resampling without replacement from the initialized prediction ensemble.
     Full years of the first lead present from the initialized are relabeled to a different year.
     """
     if (init.init.dt.year.groupby("init.year").count().diff("year") != 0).any():
         raise ValueError(
             f'`resample_uninitialized_from_initialized` only works if the same number of initializations is present each year, found {init.init.dt.year.groupby("init.year").count()}'
         )
-    # init = init.isel(lead=0, drop=True)
-    if "valid_time" in init.coords:
-        init = init.drop("valid_time")
+    if "init" not in resample_dim:
+        raise ValueError(
+            f"Only resampling on `init` makes forecasts uninitialzed. Found resample_dim={resample_dim}."
+        )
+    init = init.isel(lead=0, drop=True)
     # resample init
     init_notnull = init.where(init.notnull(), drop=True)
     full_years = list(set(init_notnull.init.dt.year.values))
@@ -395,22 +397,37 @@ def resample_uninitialized_from_initialized(init, resample_dim=["init", "member"
         m = full_years.copy()
         np.random.shuffle(m)
         years_same = (np.array(m) - np.array(full_years) == 0).any()
-    ms = [str(i) for i in m]  # convert year ints to str
 
-    resampled_inits = xr.concat([init.sel(init=i).init for i in ms], "init")
+    resampled_inits = xr.concat([init.sel(init=str(i)).init for i in m], "init")
     resampled_uninit = init.sel(init=resampled_inits)
     resampled_uninit["init"] = init_notnull.sel(
         init=slice(str(full_years[0]), str(full_years[-1]))
     ).init
+    # take time dim and overwrite with sorted
+    resampled_uninit = (
+        resampled_uninit.swap_dims({"init": "valid_time"})
+        .drop("init")
+        .rename({"valid_time": "time"})
+    )
+    resampled_uninit = resampled_uninit.assign_coords(
+        time=resampled_uninit.time.sortby("time").values
+    )
+
     # resample members
     if "member" in resample_dim:
         resampled_members = np.random.randint(0, init.member.size, init.member.size)
         resampled_uninit = resampled_uninit.isel(member=resampled_members)
         resampled_uninit["member"] = init.member
-    # thats the problem!
-    # need to find a way how to get continuous time in lead frequency
-    # other idea: dont do .isel(lead=0) but then uninitialized has lead dim
-    return resampled_uninit  #.rename({"init": "time"})
+
+    from . import __version__ as version
+
+    resampled_uninit.attrs.update(
+        {
+            "description": "created by `HindcastEnsemble.generate_uninitialized()` resampling years without replacement from initialized",
+            "documentation": f"https://climpred.readthedocs.io/en/v{version}/api/climpred.classes.HindcastEnsemble.generate_uninitialized.html#climpred.classes.HindcastEnsemble.generate_uninitialized",
+        }
+    )
+    return resampled_uninit
 
 
 def _bootstrap_by_stacking(init_pm, control):